백준 크리보드 [11058] - 3차 풀이 완료 - 정답 틀림

+++ CodingTest/Baekjoon/Main.java
@@ -16,7 +16,7 @@ public class Main {
 }
 
 class BjSolution {
-    int answer, gClickNum, gBufferCount, gStandardCount;
+    long answer, gClickNum, gBufferCount, gStandardCount;
     boolean gCopyFlag;
     boolean gSelectFlag;
